Small refactor before we go into the weeds
