+++ src/coguard_cli/util.py
@@ -7,8 +7,10 @@ import sys
 import json
 import logging
 import pathlib
+import tempfile
 from enum import Enum
 import textwrap
+from zipfile import ZipFile
 
 import shutil
 from typing import Set, Dict, Optional, Tuple, List
@@ -379,3 +381,108 @@ def extract_reference_string(entry_dict: Dict):
         config_file_in_entry.get("fileName", ".")
     )
     return f" (affected files: {reference_path}{from_to_line})"
+
+def upload_and_fix_zip_candidate(
+        zip_candidate: Optional[Tuple[str, Dict]],
+        folder_path: str,
+        token: Token,
+        coguard_api_url: str,
+        organization: Optional[str]) -> None:
+    """
+    The common function to upload and a zip file, as generated by the
+    helper functions.
+    """
+    if zip_candidate is None:
+        print(
+            f"{COLOR_YELLOW}Unable to identify any known configuration files.{COLOR_TERMINATION}"
+        )
+        return
+    zip_file, zip_manifest = zip_candidate
+    api_result = api_connection.send_zip_file_for_fixing(
+        zip_file,
+        token,
+        coguard_api_url,
+        organization
+    )
+    os.remove(zip_file)
+    if api_result is None:
+        print(f"{COLOR_RED} There was an error uploading the zip candidate.{COLOR_TERMINATION}")
+        return
+    print(f'{COLOR_CYAN} Applying the changes.{COLOR_TERMINATION}')
+    temp_folder = tempfile.mkdtemp(prefix="coguard_cli_fix_extract")
+    with ZipFile(api_result, 'r') as zip_stream:
+        zip_stream.extractall(temp_folder)
+    os.remove(api_result)
+    apply_fixes_to_folder(temp_folder, folder_path, zip_manifest)
+    print(f'{COLOR_CYAN} Done applying the changes. {COLOR_TERMINATION}')
+
+def apply_fixes_to_folder(fix_folder: str, target_folder: str, zip_manifest: Dict):
+    """
+    The helper function to apply the fixes found in fix_folder to the target folder.
+
+    It identifies all configuration files, tries to map them in the target folder
+    using the manifest information, and then performs the move.
+
+    If everything weng according to plan, the fix_folder will be deleted. Otherwise,
+    it will be left for the purpose of review.
+    """
+    files_to_move = []
+    different_services_objects = []
+    for service_name, service in zip_manifest.get('clusterServices', {}).items():
+        different_services_objects.append((os.path.join('clusterServices', service_name), service))
+    for machine_name, machine in zip_manifest.get('machines', {}).items():
+        for service_name, service in machine.get('services', {}).items():
+            different_services_objects.append(
+                (os.path.join(machine_name, service_name), service)
+            )
+    for pth, service in different_services_objects:
+        for config_file in service.get("configFileList", []):
+            files_to_move.append(
+                (
+                    os.path.join(
+                        pth,
+                        convert_posix_path_to_os_path(config_file.get("subPath")),
+                        config_file.get("fileName")
+                    ),
+                    os.path.join(
+                        convert_posix_path_to_os_path(config_file.get("subPath")),
+                        config_file.get("fileName")
+                    )
+                )
+            )
+        for config_file in service.get("complimentaryFileList", []):
+            files_to_move.append(
+                (
+                    os.path.join(
+                        pth,
+                        convert_posix_path_to_os_path(config_file.get("subPath")),
+                        config_file.get("fileName")
+                    ),
+                    os.path.join(
+                        convert_posix_path_to_os_path(config_file.get("subPath")),
+                        config_file.get("fileName")
+                    )
+                )
+            )
+    all_files_found = True
+    for file_pth_fix, file_pth_folder in files_to_move:
+        if not os.path.exists(os.path.join(target_folder, file_pth_folder)):
+            logging.error("File %s did not exist in %s", file_pth_folder, target_folder)
+            all_files_found = False
+            continue
+        try:
+            shutil.copyfile(
+                os.path.join(fix_folder, file_pth_fix),
+                os.path.join(target_folder, file_pth_folder)
+            )
+        except OSError as err:
+            all_files_found = False
+            logging.error("Could not copy %s to %s: %s",
+                          file_pth_fix,
+                          file_pth_folder,
+                          err)
+    if not all_files_found:
+        print(f"{COLOR_RED} Not all files were possible to be fixed.")
+        print(f"You can review the extracted and fixed files at {fix_folder}.{COLOR_TERMINATION}")
+    else:
+        shutil.rmtree(fix_folder)
